refactor(vault): split persistence responsibilities
